+++ app/(auth)/signin/page.tsx
@@ -1,8 +1,5 @@
 "use client";
 
-import { FormEvent, useState } from "react";
-import { Input } from "@heroui/input";
-import { Button } from "@heroui/button";
 import Image from "next/image";
 import { Card } from "@heroui/card";
 import { Divider } from "@heroui/divider";
@@ -11,24 +8,9 @@ import NextLink from "next/link";
 
 import SlackShotsLogo from "@/public/SSLOGO_NOBG.png";
 import SlackLogo from "@/public/SLA-appIcon-desktop.png";
-import { useAuth } from "@/hooks/use-auth";
+import SignInForm from "@/components/auth/sign-in-form";
 
 export default function SignInPage() {
-  const { signIn, loading, error } = useAuth();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const disabled = () => loading || !email || !password;
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    await signIn({
-      email,
-      password,
-      callbackURL: "/dashboard",
-    });
-  };
-
   return (
     <Card className="w-full max-w-md rounded-xl drop-shadow-lg p-8 gap-1 text-foreground font-semibold border border-zinc-400/25 group relative flex shadow-[inset_0_-8px_10px_#8fdfff1f] whitespace-nowrap">
       <Image
@@ -59,36 +41,7 @@ export default function SignInPage() {
         <Divider className="flex-1" />
       </div>
 
-      <form className="space-y-4" onSubmit={handleSubmit}>
-        <Input
-          isRequired
-          label="Email"
-          type="email"
-          value={email}
-          variant="faded"
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <Input
-          isRequired
-          label="Password"
-          type="password"
-          value={password}
-          variant="faded"
-          onChange={(e) => setPassword(e.target.value)}
-        />
-
-        {error && <p className="text-sm text-red-500">{error.message}</p>}
-
-        <Button
-          className="w-full rounded-lg px-4 py-3 text-white font-semibold disabled:opacity-50"
-          color={disabled() ? "default" : "primary"}
-          isDisabled={disabled()}
-          isLoading={loading}
-          type="submit"
-        >
-          {loading ? "Signing in..." : "Sign In"}
-        </Button>
-      </form>
+      <SignInForm />
 
       <p className="mt-6 text-center text-sm text-zinc-500">
         Don’t have an account?{" "}
