+++ app/(auth)/signup/page.tsx
@@ -1,8 +1,3 @@
-"use client";
-
-import { FormEvent, useState } from "react";
-import { Input } from "@heroui/input";
-import { Button } from "@heroui/button";
 import Image from "next/image";
 import { Card } from "@heroui/card";
 import { Divider } from "@heroui/divider";
@@ -11,42 +6,9 @@ import NextLink from "next/link";
 
 import SlackShotsLogo from "@/public/SSLOGO_NOBG.png";
 import SlackLogo from "@/public/SLA-appIcon-desktop.png";
-import { useAuth } from "@/hooks/use-auth";
+import SignUpForm from "@/components/auth/sign-up-form";
 
 export default function SignUpPage() {
-  const { signUp, loading, error } = useAuth();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-
-  const disabled = () => {
-    return (
-      loading ||
-      !email ||
-      !password ||
-      !confirmPassword ||
-      !firstName ||
-      !lastName
-    );
-  };
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    if (password !== confirmPassword) {
-      alert("Passwords do not match!");
-
-      return;
-    }
-    await signUp({
-      email,
-      password,
-      name: `${firstName} ${lastName}`,
-      callbackURL: "/dashboard",
-    });
-  };
-
   return (
     <Card className="w-full max-w-md rounded-xl drop-shadow-lg p-8 gap-1 text-foreground font-semibold border border-zinc-400/25 group relative flex shadow-[inset_0_-8px_10px_#8fdfff1f] whitespace-nowrap">
       <Image
@@ -77,59 +39,7 @@ export default function SignUpPage() {
         <Divider className="flex-1" />
       </div>
 
-      <form className="space-y-4" onSubmit={handleSubmit}>
-        <div className="flex gap-4">
-          <Input
-            isRequired
-            label="First Name"
-            value={firstName}
-            variant="faded"
-            onChange={(e) => setFirstName(e.target.value)}
-          />
-          <Input
-            isRequired
-            label="Last Name"
-            value={lastName}
-            variant="faded"
-            onChange={(e) => setLastName(e.target.value)}
-          />
-        </div>
-        <Input
-          isRequired
-          label="Email"
-          type="email"
-          value={email}
-          variant="faded"
-          onChange={(e) => setEmail(e.target.value)}
-        />
-        <Input
-          isRequired
-          label="Password"
-          type="password"
-          value={password}
-          variant="faded"
-          onChange={(e) => setPassword(e.target.value)}
-        />
-        <Input
-          isRequired
-          label="Confirm Password"
-          type="password"
-          value={confirmPassword}
-          variant="faded"
-          onChange={(e) => setConfirmPassword(e.target.value)}
-        />
-        {error && <p className="text-sm text-red-500">{error.message}</p>}
-
-        <Button
-          className="w-full rounded-lg px-4 py-3 text-white font-semibold disabled:opacity-50"
-          color={disabled() ? "default" : "primary"}
-          isDisabled={disabled()}
-          isLoading={loading}
-          type="submit"
-        >
-          {loading ? "Creating account..." : "Sign Up"}
-        </Button>
-      </form>
+      <SignUpForm />
 
       <p className="mt-6 text-center text-sm text-zinc-500">
         Already have an account?{" "}
